+++ tsconfig.json
@@ -1,6 +1,6 @@
 {
   // "include": ["src/**/*", "bin/**/*"],
-  "exclude": ["node_modules", "**/*.test.*", "dist"],
+  "exclude": ["node_modules", "dist"],
   "compilerOptions": {
     /* Visit https://aka.ms/tsconfig to read more about this file */
